requirements.txt
Flask>=3.0
flask-session>=0.8
spotipy>=2.24
ytmusicapi>=1.8
python-dotenv>=1.0
requests>=2.31
